requirements.txt
google-genai
customtkinter
speechrecognition
pygame
scipy
pydub
pyautogui
pocket-tts
python-dotenv